+++ Source/BE/src/main/java/imitori/neo4j/dto/WordDto.java
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class WordDto {
 
     @JsonProperty("id")
-    private Long id;
+    private Integer id;
 
     @JsonProperty("word")
     public String word;
@@ -18,15 +18,15 @@ public class WordDto {
     @JsonProperty("lang")
     public String lang;
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public WordDto(Long id, String word, ArrayList<String> spell, String lang) {
+    public WordDto(Integer id, String word, ArrayList<String> spell, String lang) {
         this.id = id;
         this.word = word;
         this.spell = spell;
