+++ Source/BE/src/main/java/imitori/neo4j/entity/JAWordNeoEntity.java
@@ -17,14 +17,14 @@ import org.neo4j.ogm.annotation.Relationship;
 public class JAWordNeoEntity {
     @Id
     @GeneratedValue
-    private Long id;
+    private Integer id;
     public String word;
     public String furigana;
     public String romaji;
     public String tl;
     
     //Depend on "lang", the id should query from JA/EN/ or even VN mongoDict
-    public Long mongoId;
+    public Integer mongoId;
 
     @Relationship(type = "SIMILAR_TO", direction = Relationship.OUTGOING)
     private ArrayList<SimilarToRelNeoEntity> similarTo = new ArrayList<>();
@@ -66,7 +66,7 @@ public class JAWordNeoEntity {
 
     }
 
-    public JAWordNeoEntity(Long id, String word, String furi, String tl) {
+    public JAWordNeoEntity(Integer id, String word, String furi, String tl) {
         this.id = id;
         this.word = word;
         this.furigana = furi;
@@ -74,7 +74,7 @@ public class JAWordNeoEntity {
         this.tl = tl;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
